+++ server/main.py
@@ -1,5 +1,5 @@
 from shared import app
-from modules.bot.bot import game_blueprint
+from server.modules.bot.api import game_blueprint
 
 app.register_blueprint(game_blueprint, url_prefix="/game")
 if __name__ == "__main__":
